+++ src/audio/AL/alutil.h
@@ -12,7 +12,7 @@
 #endif
 
 #include <glm/glm.hpp>
-#include "../typedefs.h"
+#include "../../typedefs.h"
 
 #define AL_CHECK(STATEMENT) STATEMENT; AL::check_errors(__FILE__, __LINE__)
 #define AL_GET_ERROR() AL::check_errors(__FILE__, __LINE__)
